# docs/openapi/watchdog.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Watchdog",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/watchdog/approval-inbox": {
      "get": {
        "tags": [
          "session-spine"
        ],
        "summary": "List pending approvals across projects via stable contract",
        "description": "Stable read surface for a cross-project approval inbox. Returns a versioned ReplyModel carrying ApprovalProjection rows instead of the legacy raw approvals payload.",
        "operationId": "get_approval_inbox_api_v1_watchdog_approval_inbox_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Approval Inbox Api V1 Watchdog Approval Inbox Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/watchdog/sessions": {
      "get": {
        "tags": [
          "session-spine"
        ],
        "summary": "List stable session directory",
        "description": "Stable read surface for cross-project session discovery. Returns a versioned ReplyModel carrying SessionProjection rows instead of the raw Codex runtime service task list.",
        "operationId": "list_sessions_api_v1_watchdog_sessions_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Sessions Api V1 Watchdog Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/watchdog/sessions/by-native-thread/{native_thread_id}": {
      "get": {
        "tags": [
          "session-spine"
        ],
        "summary": "Resolve stable session projection by native thread",
        "description": "Stable read surface for callers that only know the native thread_id. Returns the canonical ReplyModel carrying SessionProjection and FactRecord data.",
        "operationId": "get_session_by_native_thread_api_v1_watchdog_sessions_by_native_thread__native_thread_id__get",
        "parameters": [
          {
            "name": "native_thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Native Thread Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Session By Native Thread Api V1 Watchdog Sessions By Native Thread  Native Thread Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/watchdog/sessions/{project_id}": {
      "get": {
        "tags": [
          "session-spine"
        ],
        "summary": "Get stable session projection",
        "description": "Stable read surface for Feishu and other callers. Returns a versioned ReplyModel carrying SessionProjection and FactRecord data.",
        "operationId": "get_session_api_v1_watchdog_sessions__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Session Api V1 Watchdog Sessions  Project Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/watchdog/sessions/{project_id}/progress": {
      "get": {
        "tags": [
          "session-spine"
        ],
        "summary": "Get stable progress view",
        "description": "Stable read surface for progress-oriented intents. Returns a versioned ReplyModel carrying TaskProgressView and supporting facts.",
        "operationId": "get_progress_api_v1_watchdog_sessions__project_id__progress_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Progress Api V1 Watchdog Sessions  Project Id  Progress Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/watchdog/sessions/{project_id}/facts": {
      "get": {
        "tags": [
          "session-spine"
        ],
        "summary": "Get stable session facts truth source",
        "description": "Canonical stable facts read surface for Feishu and other callers. Returns a versioned ReplyModel carrying FactRecord rows without changing the explanation surfaces.",
        "operationId": "get_session_facts_api_v1_watchdog_sessions__project_id__facts_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Session Facts Api V1 Watchdog Sessions  Project Id  Facts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}